test: fix suite failures under isolated/parallel runs
Two pre-existing failures that surface once tests run in separate
processes (parallel) or in isolation:

- createSalePoint() was defined in PosSessionTest but used by
  PosFavoritesTest, so it was undefined whenever the two didn't share a
  process. Move it to tests/Pest.php (global helpers) and drop the
  duplicate definition.
- StockMovementLedgerHardeningTest's adjustment test did an upward
  adjustment under the default purchase_driven strategy, which M4's gate
  now blocks (the test predates the gate and was never re-run against it).
  Set free_form for that case, mirroring StockAdjustmentTest.

Full suite now green under --parallel (3520 assertions, 0 failed).
